[Ref] Separate class for mzSlice #237
Added a new directory - "datastructures". <datastructures> will store
custom datatypes such as mzSample and mzSlice

remove class <mzSlice> from "mzSample.h"

Issue: #237
